refac: Move help output into own module
